% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_dark_blue.R
\name{district_dark_blue}
\alias{district_dark_blue}
\title{District Dark Blue}
\usage{
district_dark_blue()
}
\value{
String character with hex color code
}
\description{
This function stores the hex code for district dark blue
}
\section{Last Updated By}{
 Stevens Dormezil
}

\section{Last Updated}{
 6/14/2022
}

\examples{
district_dark_blue()

}
