% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_light_blue.R
\name{district_light_blue}
\alias{district_light_blue}
\title{District Light Blue}
\usage{
district_light_blue()
}
\description{
This function stores the hex code for district dark blue
}
\examples{
district_light_blue()
}
