% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_red.R
\name{district_red}
\alias{district_red}
\title{District Red}
\usage{
district_red()
}
\value{
String character with hex color code
}
\description{
This function stores the hex code for district red
}
\section{Last Updated By}{
 Stevens Dormezil
}

\section{Last Updated}{
 6/14/2022
}

\examples{
district_red()

}
