% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_yellow.R
\name{district_yellow}
\alias{district_yellow}
\title{District Yellow}
\usage{
district_yellow()
}
\value{
String character with hex color code
}
\description{
This function stores the hex code for district yellow
}
\section{Last Updated By}{
 Stevens Dormezil
}

\section{Last Updated}{
 6/14/2022
}

\examples{
district_yellow()

}
