% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreaformat-package.R
\docType{package}
\name{dreaformat-package}
\alias{dreaformat}
\alias{dreaformat-package}
\title{dreaformat: DREA formating tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains a collection of tools for formating tables and plots according to department standards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dormezil/dreaformat}
  \item Report bugs at \url{https://github.com/dormezil/dreaformat/issues}
}

}
\author{
\strong{Maintainer}: Stevens Dormezil \email{stevens.dormezil@palmbeachschools.org} (\href{https://orcid.org/0000-0003-1886-6964}{ORCID})

}
\keyword{internal}
