% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbsc_green.R
\name{pbsc_green}
\alias{pbsc_green}
\title{Palm Beach State College Green}
\usage{
pbsc_green()
}
\value{
String of hex code
}
\description{
This function stores the hex code for PBSC
}
\examples{
pbsc_green()
}
