% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drea_table.R
\name{drea_table}
\alias{drea_table}
\title{Table formatting}
\usage{
drea_table(dataframe)
}
\arguments{
\item{dataframe}{Input dataframe to be converted to flextable according to department standards}
}
\description{
This function formats tables into drea standard for importing into word documents
}
